fix: add datadog service name during initialization
